Generate options for optimization

genopts() initializes and generates core options and settings for modeling and optimization, including random effects, simulation settings, and likelihood approximations.

Usage

genopts(
  f,
  g,
  xt,
  p,
  h,
  nsim = 1,
  n = 30,
  adist = NULL,
  interact = TRUE,
  fo_appr = (nsim < 10),
  biseq = NA,
  omega_expansion = 1,
  p_thetai = NA,
  single_betas = NA
)

Arguments

f

The likelihood function

g

The model function

xt

The data

p

The parameter settings

h

The error function

nsim

The number of simulations

n

The number of simulations

adist

The distribution of the random effects

interact

Whether to include interaction terms

fo_appr

Whether to use first-order approximation

biseq

The sequence of random effects

omega_expansion

The expansion factor for the omega matrix

p_thetai

The theta matrix

single_betas

The beta matrix

Value

A fitted model

Examples

#test