Fitting aggregate data

fitIRMC() implements the Iterative Reweighting (IRMC) algorithm for parameter estimation of the given aggregate data model, iterating over maximum likelihood updates with weighted MC updates. This version of the function uses nloptr instead of optimx. TOL = 1e-10

Usage

fitIRMC(
  opts,
  obs,
  maxiter = 100,
  convcrit_nll = 1e-05,
  single_dataframe = TRUE,
  phase_fractions = c(0.2, 0.4, 0.2, 0.2),
  max_worse_iterations = 10,
  chains = 1,
  pertubation = 0.1,
  seed = 1
)

Arguments

opts

options

obs

observed data

maxiter

maximum number of iterations

convcrit_nll

convergence criterion for the negative log-likelihood

single_dataframe

single model or multiple models

phase_fractions

vector of phase fractions

max_worse_iterations

maximum number of consecutive worse iterations before skipping a phase

chains

number of chains

pertubation

pertubation factor for the initial parameter values of each chain

seed

random seed

Value

A fitted model

Examples

#test