Fitting aggregate data

timedEM() implements the Expectation-Maximization(EM) algorithm for parameter estimation of the given aggregate data model, iterating over maximum likelihood updates with weighted MC updates. It is used to compare the performance of the old and new implementation of aggregate data modelling.

Usage

timedEM(init, opts, obs, maxiter = 100, convcrit_nll = 5e-04, nomap = TRUE)

Arguments

init

initial parameter values

opts

options

obs

observed data

maxiter

maximum number of iterations

convcrit_nll

convergence criterion for the negative log-likelihood

nomap

single model or multiple models

Value

A fitted model

Examples

#test